requirements.txt
flask==2.3.3
opencv-python>=4.5.0
numpy>=1.19.0
pupil-apriltags>=1.0.4
pyserial>=3.5